Created work time resource and moved its functionalities from the place resource to it. Injected the user service into the place service.
