+++ src/place/entities/place.entity.ts
@@ -11,9 +11,9 @@ import {
   PrimaryGeneratedColumn,
   UpdateDateColumn,
 } from 'typeorm';
-import { WorkTime } from './work-time.entity';
 import { SocialMedias } from './social-medias.entity';
 import { User } from 'src/user/entities/user.entity';
+import { WorkTime } from 'src/work-time/entities/work-time.entity';
 
 @Entity()
 export class Place {
